Move parseEdgeId logic
